fix: cluster U — API design tightening (2 findings, #1463)
Two non-breaking API hygiene findings:

API-V1.38-7 — `ExportModel.dim: Option<u64>` → `Option<usize>`. Every
other dim field in the codebase is `usize` (`ModelConfig.dim`,
`EmbeddingConfig.dim`, `VectorIndex::dim()`, `Embedding::len()`); the
u64 outlier required `as usize` casts at the handler boundary.
`--dim 768` parses identically. Test suite passed unchanged after
re-typing `dim_override: Option<u64>` → `Option<usize>` in
`cmd_export_model` + `resolve_dim` + `write_model_toml`.

API-V1.38-8 — `cqs status --wait-secs` and `cqs eval --require-fresh-secs`
have identical semantics (block until daemon reports `state == fresh`,
bounded). Two flag spellings = muscle-memory cost — agents that learn
one don't transfer. Each now exposes the other as a `visible_alias`
so both paths parse cleanly. Defaults stay command-specific (30 for
status, 600 for eval). Smoke-tested both directions:
- `cqs status --require-fresh-secs 5` parses ✓
- `cqs eval --wait-secs 5 ...` parses ✓

Skipped from the original cluster scope (too broad for one PR):
- API-V1.38-1 (ModelCommand/HookCommand → TextJsonArgs envelope) —
  separate cluster
- API-V1.38-2 (ProjectCommand::Search flatten SearchArgs) — separate
- API-V1.38-9 (HfHub error variants → ResolveError collapse) — breaking,
  separate
- API-V1.38-10 (LimitArg fan-out across 7 sister args) — 61 caller
  sites; large mechanical change, separate cluster

Verified: cargo build, cargo test --bin cqs export_model (11/11),
cargo clippy -- -D warnings, cargo fmt --check. Smoke-tested alias
resolution for `--wait-secs` ↔ `--require-fresh-secs`.

+++ src/cli/definitions.rs
@@ -783,8 +783,14 @@ pub(super) enum Commands {
         #[arg(long, default_value = ".")]
         output: std::path::PathBuf,
         /// Embedding dimension override (auto-detected from config.json if omitted)
+        ///
+        /// API-V1.38-7 (#1463): `usize` aligns with `ModelConfig.dim`,
+        /// `EmbeddingConfig.dim`, `VectorIndex::dim()` etc. — every other
+        /// dim field in the codebase. The previous `Option<u64>` parsed
+        /// identically (`--dim 768`) but required `as usize` casts at the
+        /// handler boundary.
         #[arg(long)]
-        dim: Option<u64>,
+        dim: Option<usize>,
     },
     /// Generate training data for fine-tuning from git history
     #[cqs_cmd(group = "a", batch = "cli")]
@@ -894,7 +900,12 @@ pub(super) enum Commands {
         ///
         /// Note: `cqs eval --require-fresh-secs` has the same semantics;
         /// default differs by use case (eval default = 600, status = 30).
-        #[arg(long, default_value_t = 30)]
+        ///
+        /// API-V1.38-8 (#1463): `--require-fresh-secs` exposed as a
+        /// visible alias so an agent that learned the eval-side spelling
+        /// can use it on `cqs status` too. Both spellings parse to the
+        /// same field; prefer `--wait-secs` in CI scripts (shorter).
+        #[arg(long, visible_alias = "require-fresh-secs", default_value_t = 30)]
         wait_secs: u64,
         /// Output as JSON. Without this, prints a one-line human summary.
         ///
